stable version

+++ CyberBall/inc/tim2.h
@@ -8,5 +8,5 @@
 #include "usart.h"
 
 void init_tim2();
-extern uint16_t pos[3];
+extern int pos[3];
 #endif
